mexc_bot/requirements.txt
numpy>=1.26,<2.0
pandas>=2.2,<3.0
httpx>=0.27,<1.0
websockets>=12.0,<13.0
tenacity>=8.2,<9.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
loguru>=0.7,<1.0
python-telegram-bot>=21.0,<22.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
uvicorn[standard]>=0.30,<1.0
pytest>=8.2,<9.0
matplotlib>=3.9,<4.0     # нужна только, если захотите локально строить графики
